+++ src/CHANGES.md
@@ -1,5 +1,9 @@
 # Development Changes
 
+## 0.8.73 - 2024-02-03
+* fix nullpointer during communication #1401
+* added `max_power` to MqTT total values #1375
+
 ## 0.8.72 - 2024-02-03
 * fixed translation #1403
 * fixed sending commands to inverters which are soft turned off #1397